requirements.txt
openai>=1.0.0
google-generativeai>=0.3.0
qdrant-client>=1.7.0
numpy>=1.24.0
fastapi>=0.104.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
uvicorn>=0.24.0

